+++ packages/content/api/api-surface.d.ts
@@ -19,3 +19,4 @@ export * from '../media';
 export * from '../personalize';
 export * from '../site';
 export * from '../tools';
+export * from '../node-tools';