+++ packages/content/src/tools/generateSites.test.ts
@@ -44,7 +44,7 @@ describe('generateSites', () => {
     consoleErrorStub = sinon.stub(console, 'error');
 
     const generateSitesModule = proxyquire('./generateSites', {
-      '@sitecore-content-sdk/core/tools': { ensurePathExists: ensurePathExistsStub },
+      '@sitecore-content-sdk/core/node-tools': { ensurePathExists: ensurePathExistsStub },
     });
     generateSites = generateSitesModule.generateSites;
   });
@@ -137,3 +137,4 @@ describe('generateSites', () => {
     expect(consoleErrorStub.calledWith(chalk.red('Error fetching site information'))).to.be.true;
   });
 });
+