+++ packages/content/src/tools/generateSites.ts
@@ -1,7 +1,7 @@
 import path from 'path';
 import chalk from 'chalk';
 import fs from 'fs';
-import { ensurePathExists } from '@sitecore-content-sdk/core/tools';
+import { ensurePathExists } from '@sitecore-content-sdk/core/node-tools';
 import { SiteInfo, SiteInfoService } from '../site';
 import { SitecoreConfig } from '../config';
 import { createGraphQLClientFactory } from '../client';