+++ packages/content/src/tools/index.ts
@@ -1,13 +1,13 @@
-export { generateSites, GenerateSitesConfig } from './generateSites';
-export { scaffoldComponent } from './scaffold';
 export { GenerateMapFunction, GenerateMapArgs } from './generate-map';
-export { extractFiles } from './codegen/extract-files';
-export {
-  writeImportMap,
-  WriteImportMapArgs,
-  WriteImportMapArgsInternal,
-  defaultMapTemplate as defaultImportMapTemplate,
-  ModuleExports,
-} from './codegen/import-map';
 export { getComponentSpec, getComponentSpecUrl } from './codegen/component-generation';
-export * from './templating';
+export {
+  ComponentFile,
+  ComponentImport,
+  ComponentFileWithType,
+  ComponentType,
+  RouterType,
+  filterComponentsByType,
+  ComponentMapEntry,
+  ComponentMapTemplate,
+  EnhancedComponentMapTemplate,
+} from './templating';