+++ packages/create-content-sdk-app/src/templates/nextjs/sitecore.cli.config.ts
@@ -5,7 +5,7 @@ import {
   generateMetadata,
   extractFiles,
   writeImportMap,
-} from '@sitecore-content-sdk/nextjs/tools';
+} from '@sitecore-content-sdk/nextjs/node-tools';
 
 export default defineCliConfig({
   config: scConfig,
@@ -25,3 +25,4 @@ export default defineCliConfig({
     exclude: ['src/components/content-sdk/*'],
   },
 });
+