+++ packages/nextjs/src/editing/editing-config-middleware.ts
@@ -4,7 +4,7 @@ import {
   QUERY_PARAM_EDITING_SECRET,
 } from '@sitecore-content-sdk/content/editing';
 import debug from '../debug';
-import { Metadata } from '@sitecore-content-sdk/core/tools';
+import { Metadata } from '@sitecore-content-sdk/core/node-tools';
 import { getEnforcedCorsHeaders } from '@sitecore-content-sdk/core/tools';
 import { EditMode } from '@sitecore-content-sdk/content/layout';
 import { getEditingSecret } from '../utils/utils';
@@ -84,3 +84,4 @@ export class EditingConfigMiddleware {
     });
   };
 }
+