+++ packages/nextjs/src/route-handler/editing-config-route-handler.ts
@@ -3,7 +3,7 @@ import {
   EDITING_ALLOWED_ORIGINS,
   QUERY_PARAM_EDITING_SECRET,
 } from '@sitecore-content-sdk/content/editing';
-import { Metadata } from '@sitecore-content-sdk/core/tools';
+import { Metadata } from '@sitecore-content-sdk/core/node-tools';
 import { getEnforcedCorsHeaders } from '@sitecore-content-sdk/core/tools';
 import { EditMode } from '@sitecore-content-sdk/content/layout';
 import { getEditingSecret } from '../utils/utils';
@@ -149,3 +149,4 @@ export const createEditingConfigRouteHandler = (options: EditingConfigRouteHandl
 
   return { GET, OPTIONS };
 };
+