+++ packages/nextjs/src/tools/index.ts
@@ -1,14 +1,16 @@
 export {
-  generateSites,
-  GenerateSitesConfig,
   ComponentFile,
   ComponentImport,
-  getComponentList,
   GenerateMapArgs,
-  extractFiles,
 } from '@sitecore-content-sdk/content/tools';
+export {
+  generateSites,
+  GenerateSitesConfig,
+  getComponentList,
+  extractFiles,
+} from '@sitecore-content-sdk/content/node-tools';
 
-export { generateMetadata } from '@sitecore-content-sdk/core/tools';
+export { generateMetadata } from '@sitecore-content-sdk/core/node-tools';
 
 export { generateMap } from './generate-map';
 export { writeImportMap } from './codegen/import-map';